feat: extract clip manager service
